+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project does not yet adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 for setuptools_scm/PEP 440 reasons.
 
+## UNRELEASED
+
+### Fixed
+
+- Keep syncing if we fail to create a PoolWallet
+
 ## 1.2.3 Chia blockchain 2021-07-26
 
 ### Added